requirements.txt
Flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
Werkzeug==2.3.7
requests==2.31.0
